requirements.txt
numpy==1.19.1
torch==1.6.0
tqdm.__version__
pandas==1.1.5
pytorch_lightning==0.9.0
scikit-learn==0.20.4
pytorch_metric_learning==0.9.92
tensorflow==2.1.0
tensorboard==2.2.0
transformers==2.5.1
tokenizers==0.9.0
